+++ packages/react-magma-dom/src/components/Dropdown/DropdownButton.tsx
@@ -1,15 +1,17 @@
+import { css } from '@emotion/core';
 import * as React from 'react';
 import { IconButton, ButtonIconPosition } from '../IconButton';
 import {
   ArrowDropUpIcon,
   ArrowLeftIcon,
   ArrowRightIcon,
-  CaretDownIcon,
+  ArrowDropDownIcon,
   IconProps,
 } from 'react-magma-icons';
 import { DropdownContext, DropdownDropDirection } from '.';
 import { Omit, useForkedRef, useGenerateId, XOR } from '../../utils';
 import { ButtonProps, ButtonSize } from '../Button';
+import styled from '../../theme/styled';
 import { ThemeContext } from '../../theme/ThemeContext';
 import { ThemeInterface } from '../../theme/magma';
 
@@ -63,6 +65,21 @@ function getButtonPadding(theme: ThemeInterface, size?: ButtonSize) {
   }
 }
 
+const StyledIconButton = styled(IconButton)`
+  ${props =>
+    props.iconPosition === ButtonIconPosition.right &&
+    props.children &&
+    css`
+      padding-right: ${getButtonPadding(props.theme, props.size)};
+    `}
+  ${props =>
+    props.iconPosition === ButtonIconPosition.left &&
+    props.children &&
+    css`
+      padding-left: ${getButtonPadding(props.theme, props.size)};
+    `}
+`;
+
 export const DropdownButton = React.forwardRef<
   HTMLButtonElement,
   DropdownButtonProps
@@ -84,7 +101,7 @@ export const DropdownButton = React.forwardRef<
         return <ArrowDropUpIcon testId="caretUp" />;
 
       default:
-        return <CaretDownIcon testId="caretDown" />;
+        return <ArrowDropDownIcon testId="caretDown" />;
     }
   }
 
@@ -114,7 +131,7 @@ export const DropdownButton = React.forwardRef<
     : ButtonIconPosition.right;
 
   return (
-    <IconButton
+    <StyledIconButton
       {...other}
       aria-expanded={context.isOpen}
       aria-haspopup="true"
@@ -124,12 +141,9 @@ export const DropdownButton = React.forwardRef<
       onClick={handleClick}
       onKeyDown={context.handleButtonKeyDown}
       ref={ref}
-      style={{
-        paddingRight: getButtonPadding(theme, props.size),
-        ...props.style,
-      }}
+      theme={theme}
     >
       {children}
-    </IconButton>
+    </StyledIconButton>
   );
 });