+++ packages/react-magma-dom/src/index.ts
@@ -1,6 +1,7 @@
 export { Alert, AlertProps } from './components/Alert';
 export { AlertVariant } from './components/AlertBase';
 export { Announce, AnnounceProps } from './components/Announce';
+export { AppBar, AppBarProps, AppBarPosition } from './components/AppBar';
 export {
   Badge,
   BadgeColor,
@@ -77,10 +78,15 @@ export {
   DropdownButton,
   DropdownButtonProps,
 } from './components/Dropdown/DropdownButton';
+export * from './components/Flex';
 export { Form, FormProps } from './components/Form';
 export { FormGroup, FormGroupProps } from './components/FormGroup';
 export { Heading, HeadingProps } from './components/Heading';
-export { HideAtBreakpoint } from './components/HideAtBreakpoint';
+export {
+  HideAtBreakpoint,
+  HideAtBreakpointDisplayType,
+  HideAtBreakpointProps,
+} from './components/HideAtBreakpoint';
 export { Hyperlink, HyperlinkProps } from './components/Hyperlink';
 export {
   IconButton,
@@ -133,6 +139,7 @@ export {
   SkipLinkContent,
   SkipLinkContentProps,
 } from './components/SkipLinkContent';
+export { Spacer, SpacerProps, SpacerAxis } from './components/Spacer';
 export { Spinner, SpinnerProps } from './components/Spinner';
 export {
   Table,
@@ -194,7 +201,7 @@ export {
   VisuallyHiddenProps,
 } from './components/VisuallyHidden';
 
-export { magma } from './theme/magma';
+export { magma, ThemeInterface } from './theme/magma';
 export { GlobalStyles } from './theme/GlobalStyles';
 export { ThemeContext } from './theme/ThemeContext';
 export { I18nContext } from './i18n';