+++ packages/react-magma-dom/src/theme/magma.ts
@@ -123,6 +123,21 @@ export interface TypographyVisualStyles {
   bodyXSmall: VisualStyle;
 }
 
+interface AppBar {
+  backgroundColor: string;
+  height: string;
+  padding: string;
+  textColor: string;
+  compact: {
+    height: string;
+    padding: string;
+  };
+  inverse: {
+    backgroundColor: string;
+    textColor: string;
+  };
+}
+
 export interface Modal {
   width: {
     small: string;
@@ -162,6 +177,7 @@ export interface Tooltip {
 }
 
 export interface ThemeInterface {
+  appBar: AppBar;
   breakpoints: Breakpoints;
   bodyFont: string;
   bodyExpressiveFont: string;
@@ -173,6 +189,7 @@ export interface ThemeInterface {
   dropdown: Dropdown;
   headingFont: string;
   iconSizes: IconSizes;
+  spacingMultiplier: number;
   spaceScale: SpacingScale;
   headingExpressiveFont: string;
   headingNarrativeFont: string;
@@ -261,7 +278,7 @@ const colors = {
   dangerInverse: '#F59295',
   neutral: '#3F3F3F', // main dark grey text color
   neutral02: '#575757',
-  neutral03: '#727272',
+  neutral03: '#707070',
   neutral04: '#8f8f8f', // lightest gray that meets 3:1 contrast ratio
   neutral05: '#BFBFBF',
   neutral06: '#DFDFDF',
@@ -281,6 +298,23 @@ const colors = {
   tone02: 'rgba(63,63,63,0.1)',
 };
 
+const spaceScale = {
+  spacing01: '2px',
+  spacing02: '4px',
+  spacing03: '8px',
+  spacing04: '12px',
+  spacing05: '16px',
+  spacing06: '24px',
+  spacing07: '28px',
+  spacing08: '32px',
+  spacing09: '40px',
+  spacing10: '48px',
+  spacing11: '56px',
+  spacing12: '64px',
+  spacing13: '96px',
+  spacing14: '160px',
+};
+
 export const magma = {
   bodyFont: '"Open Sans",Helvetica,sans-serif',
   bodyExpressiveFont: '"Open Sans",Helvetica,sans-serif',
@@ -291,6 +325,7 @@ export const magma = {
   headingExpressiveFont: '"Open Sans",Helvetica,sans-serif',
   headingNarrativeFont: "'Noto Serif',Times New Roman,serif",
   direction: 'ltr',
+  spacingMultiplier: 8,
 
   // breakpoints
   breakpoints: {
@@ -301,22 +336,7 @@ export const magma = {
     xl: 1200,
   },
 
-  spaceScale: {
-    spacing01: '2px',
-    spacing02: '4px',
-    spacing03: '8px',
-    spacing04: '12px',
-    spacing05: '16px',
-    spacing06: '24px',
-    spacing07: '28px',
-    spacing08: '32px',
-    spacing09: '40px',
-    spacing10: '48px',
-    spacing11: '56px',
-    spacing12: '64px',
-    spacing13: '96px',
-    spacing14: '160px',
-  },
+  spaceScale: spaceScale,
 
   iconSizes: {
     xSmall: 16,
@@ -472,9 +492,18 @@ export const magma = {
     },
   },
 
-  dropdown: {
-    content: {
-      maxHeight: '250px',
+  appBar: {
+    backgroundColor: colors.neutral08,
+    height: '88px',
+    padding: `${spaceScale.spacing06} ${spaceScale.spacing05}`,
+    textColor: colors.neutral,
+    compact: {
+      height: '56px',
+      padding: `${spaceScale.spacing05} ${spaceScale.spacing05} ${spaceScale.spacing05} ${spaceScale.spacing06}`,
+    },
+    inverse: {
+      backgroundColor: colors.foundation02,
+      textColor: colors.neutral08,
     },
   },
 
@@ -484,8 +513,8 @@ export const magma = {
     },
   },
 
-  select: {
-    menu: {
+  dropdown: {
+    content: {
       maxHeight: '250px',
     },
   },
@@ -495,6 +524,12 @@ export const magma = {
       small: '300px',
       medium: '600px',
       large: '900px',
+    },
+  },
+
+  select: {
+    menu: {
+      maxHeight: '250px',
     },
   },